+++ frontend/src/api/Organisation.js
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 import Service from './HttpCommon'
+import URLSearchParams from '@ungap/url-search-params'
 
 export default {
 
@@ -43,9 +44,22 @@ export default {
    * @param {Date} atDate - Date
    * @returns {Array} List of organisation units within the organisation
    */
-  getChildren (uuid, atDate) {
+  getChildren (uuid, atDate, extra) {
     if (atDate instanceof Date) atDate = atDate.toISOString().split('T')[0]
-    return Service.get(`/o/${uuid}/children?at=${atDate}`)
+
+    const params = new URLSearchParams()
+
+    if (atDate) {
+      params.append('at', atDate)
+    }
+
+    if (extra !== undefined) {
+      for (const key in extra) {
+        params.append(key, extra[key])
+      }
+    }
+
+    return Service.get(`/o/${uuid}/children?${params}`)
       .then(response => {
         return response.data
       })